fix(codegen): [STABLE] defined wrap for signed integer arithmetic via unsigned cast (closes #56)
- operators.js: binary +,-,* emit (intN_t)((uintN_t)a OP (uintN_t)b) for signed types
- assign.js: compound +=,-=,*= emit same pattern; widening check runs BEFORE cast
- operators.js: / and % get INT_MIN/-1 overflow guard (from #54)
- assign.js: /= and %= get same guard
- 87 expected.c files regenerated to match new output
- All 1676 tests pass

+++ test/cases/phase1/chain/new-class-chain/expected.c
@@ -11,7 +11,7 @@ static Pair Pair_new(int32_t a, int32_t b) {
 }
 
 static int32_t Pair_sum(const Pair *self) {
-    return self->a + self->b;
+    return (int32_t)((uint32_t)self->a + (uint32_t)self->b);
 }
 
 static Wrapper Wrapper_new(int32_t x, int32_t y) {
