+++ test/cases/phase1/functions/recursive/expected.c
@@ -4,7 +4,7 @@ int32_t factorial_i32(int32_t n) {
     if (n <= 1) {
         return 1;
     }
-    return n * factorial_i32(n - 1);
+    return (int32_t)((uint32_t)n * (uint32_t)factorial_i32(n - 1));
 }
 
 int main(void) {
