+++ test/cases/phase11/ram-stack/stack-callchain-ok/expected.c
@@ -3,17 +3,17 @@
 int32_t leaf(void) {
     int32_t a = 1;
     int32_t b = 2;
-    return a + b;
+    return (int32_t)((uint32_t)a + (uint32_t)b);
 }
 
 int32_t mid(void) {
     int32_t x = 3;
-    return x + leaf();
+    return (int32_t)((uint32_t)x + (uint32_t)leaf());
 }
 
 int32_t top(void) {
     int32_t y = 4;
-    return y + mid();
+    return (int32_t)((uint32_t)y + (uint32_t)mid());
 }
 
 int main(void) {
