+++ test/cases/phase2/truthy/while-opt/expected.c
@@ -9,7 +9,7 @@ int main(void) {
     int32_t i = 0;
     while (i < 3) {
         if (x.has_value && x.value != 0) {
-            sum = sum + x.value;
+            sum = (int32_t)((uint32_t)sum + (uint32_t)x.value);
         }
         i = i + 1;
     }
