+++ test/cases/phase2/utility-types/parameters-type/expected.c
@@ -3,12 +3,12 @@
 typedef struct { int32_t _0; int32_t _1; } tuple_i32_i32;
 
 int32_t add_i32_i32(int32_t a, int32_t b) {
-    return a + b;
+    return (int32_t)((uint32_t)a + (uint32_t)b);
 }
 
 int main(void) {
     TSC_INIT();
     const tuple_i32_i32 args = {._0 = 3, ._1 = 4};
-    printf("%d\n", args._0 + args._1);
+    printf("%d\n", (int32_t)((uint32_t)args._0 + (uint32_t)args._1));
     return 0;
 }
