+++ test/cases/phase7/async-forof/forof-continue-await/expected.c
@@ -40,7 +40,7 @@ case_1:
             if (!self->_await_0._done) return;
             int32_t v = self->_await_0._result;
             if (v < 0) goto forof_1_cont;
-            self->total = self->total + v;
+            self->total = (int32_t)((uint32_t)self->total + (uint32_t)v);
 forof_1_cont:
             self->_forof_idx_0++;
             if (!(self->_forof_idx_0 < self->arr.length)) { goto forof_1_end; }
