+++ test/cases/phase9/strict/ok-clean-code/expected.c
@@ -1,7 +1,7 @@
 #include "runtime.h"
 
 int32_t add_i32_i32(int32_t a, int32_t b) {
-    return a + b;
+    return (int32_t)((uint32_t)a + (uint32_t)b);
 }
 
 int main(void) {
